split schemas to auth and data
